Use async processing to avoid database timeouts
In the previous handling of events, event were handled
immediately during the transaction, in the same thread. This caused a
lot of delays, blocking the database threads. The EventListener annotations are replaced by
TransactionalEventListener to ensure that the previous transaction has
been committed before processing the event. It explicitly creates
a new transaction to avoid overlapping transactions. Finally, it enables
async processing to avoid the existing transaction threads being blocked.